refactor(archtest): 迁 Saga 规则族 → module-path-agnostic CellRule [#1634]
M3 PR-4: migrate the 13 SAGA-* archtest invariants out of the 5968-line
saga_invariants_test.go into an importable non-test home, deriving every
platform-symbol path from PlatformModulePath (no bare literals), and ratchet
the ARCHTEST-MODULE-PATH-FUNNEL-01 baseline down by one line.

- New saga_invariants.go: 11 importable Check* funcs (bucket-1 COMPENSATE-PURE
  + bucket-2 ×10) + their scanner/helper logic, all platform paths derived from
  PlatformModulePath. SAGA-STEP-COMPENSATE-PURE-01 registered in
  StandardCellRules(); the 10 internal-layout / conformance rules stay
  importable-but-unregistered (vacuous/false-red externally), documented in the
  StandardCellRules godoc.
- saga_invariants_test.go dogfoods the same Check* (single source): single-part
  detectors call Check* directly; multi-part rules keep granular An/Bn tests +
  add an aggregate _CheckDogfood. All // INVARIANT: SAGA-* anchors + Test* stay
  here (SAGA-INVARIANTS-FILE-CONSOLIDATED-01). SAGA-STATUS-FANOUT-COVERAGE-01 +
  SAGA-INVARIANTS-FILE-CONSOLIDATED-01 (golden/meta-bound) stay test-only,
  literals derived in place.
- New shared_helpers.go: helpers needed by the non-test Check* but previously
  only in *_test.go (posInRanges, collectFuncBodyRanges, isContextType,
  isTimeDurationType, itoa, sliceOrNone) relocated from span/clock/tenant/
  healthcheck/outbox test files — single source, no duplication.
- collectSagaJournalImpls keeps the typesutil.ImplementsInterface funnel
  (TYPESUTIL-IMPLEMENTS-FUNNEL-01), not raw go/types.Implements.
- testdata/module_path_funnel.baseline: drop saga_invariants_test.go line.

Behavior-preserving: detection logic moved not rewritten; drivers/scopes and
all Hard/Medium ratings preserved; verified by the existing saga tests + RED
fixtures staying GREEN and the baseline ratchet flipping to GREEN.

Refs: Closes #1634
ref: tools/archtest/errcode_invariants.go (migration recipe, #1302/a8bb9df92)
