+++ tools/archtest/healthcheck_verify_runtime_test.go
@@ -146,28 +146,6 @@ esac
 `
 }
 
-func itoa(n int) string {
-	if n == 0 {
-		return "0"
-	}
-	neg := n < 0
-	if neg {
-		n = -n
-	}
-	var buf [20]byte
-	i := len(buf)
-	for n > 0 {
-		i--
-		buf[i] = byte('0' + n%10)
-		n /= 10
-	}
-	if neg {
-		i--
-		buf[i] = '-'
-	}
-	return string(buf[i:])
-}
-
 func indent(s string) string {
 	if s == "" {
 		return "  <empty>"
